Hubspot profile sync update
