+++ profiles/views_test.py
@@ -53,7 +53,6 @@ def test_get_user_by_id(user_client, user):
             "company_size": user.profile.company_size,
             "job_title": user.profile.job_title,
             "birth_year": int(user.profile.birth_year),
-            "leadership_level": user.profile.leadership_level,
             "job_function": user.profile.job_function,
             "years_experience": user.profile.years_experience,
             "highest_education": user.profile.highest_education,
@@ -88,7 +87,6 @@ def test_get_user_by_me(user_client, user):
             "company_size": user.profile.company_size,
             "job_title": user.profile.job_title,
             "birth_year": int(user.profile.birth_year),
-            "leadership_level": user.profile.leadership_level,
             "job_function": user.profile.job_function,
             "years_experience": user.profile.years_experience,
             "highest_education": user.profile.highest_education,
