+++ static/js/flow/authTypes.js
@@ -73,7 +73,6 @@ export type Profile = {
   job_function: ?string,
   years_experience: ?number,
   company_size: ?number,
-  leadership_level: ?string,
   highest_education: ?string
 }
