+++ static/js/lib/validation.js
@@ -1,5 +1,14 @@
 // @flow
+import { __, includes } from "ramda"
 import * as yup from "yup"
+import {
+  ADDRESS_LINES_MAX,
+  CA_ALPHA_2,
+  CA_POSTAL_CODE_REGEX,
+  COUNTRIES_REQUIRING_STATE,
+  US_ALPHA_2,
+  US_POSTAL_CODE_REGEX
+} from "../constants"
 
 // Field validations
 
@@ -53,4 +62,115 @@ export const changeEmailFormValidation = yup.object().shape({
   ),
 
   confirmPassword: passwordFieldValidation.label("Confirm Password")
+})
+
+export const legalAddressValidation = yup.object().shape({
+  profile: yup.object().shape({
+    name: yup
+      .string()
+      .label("Full Name")
+      .trim()
+      .required()
+      .min(2)
+  }),
+  legal_address: yup.object().shape({
+    first_name: yup
+      .string()
+      .label("First Name")
+      .trim()
+      .required(),
+    last_name: yup
+      .string()
+      .label("Last Name")
+      .trim()
+      .required(),
+    city: yup
+      .string()
+      .label("City")
+      .trim()
+      .required(),
+    street_address: yup
+      .array()
+      .label("Street address")
+      .of(yup.string().max(60))
+      .min(1)
+      .max(ADDRESS_LINES_MAX)
+      .compact()
+      .required(),
+    state_or_territory: yup
+      .mixed()
+      .label("State/Territory")
+      .when("country", {
+        is:   includes(__, COUNTRIES_REQUIRING_STATE),
+        then: yup.string().required()
+      }),
+    country: yup
+      .string()
+      .label("Country")
+      .length(2)
+      .required(),
+    postal_code: yup
+      .string()
+      .label("Zip/Postal Code")
+      .trim()
+      .when("country", (country, schema) => {
+        if (country === US_ALPHA_2) {
+          return schema.required().matches(US_POSTAL_CODE_REGEX, {
+            message:
+              "Postal Code must be formatted as either 'NNNNN' or 'NNNNN-NNNN'"
+          })
+        } else if (country === CA_ALPHA_2) {
+          return schema.required().matches(CA_POSTAL_CODE_REGEX, {
+            message: "Postal Code must be formatted as 'ANA NAN'"
+          })
+        }
+      })
+  })
+})
+
+export const passwordValidation = yup.object().shape({
+  password: newPasswordFieldValidation
+})
+
+export const profileValidation = yup.object().shape({
+  profile: yup.object().shape({
+    gender: yup
+      .string()
+      .label("Gender")
+      .required(),
+    birth_year: yup
+      .string()
+      .label("Birth Year")
+      .required(),
+    company: yup
+      .string()
+      .label("Company")
+      .trim()
+      .required(),
+    job_title: yup
+      .string()
+      .label("Job Title")
+      .trim()
+      .required(),
+    industry: yup
+      .string()
+      .label("Industry")
+      .required(),
+    job_function: yup
+      .string()
+      .label("Job Function")
+      .required(),
+    company_size: yup
+      .string()
+      .label("Company Size")
+      .required(),
+    years_experience: yup
+      .string()
+      .label("Years of Work Experience")
+      .required(),
+    highest_education: yup
+      .string()
+      .label("Highest Level of Education")
+      .required()
+  })
 })
